+++ src/prefabs/pageWithCrudDialogs.tsx
@@ -2194,7 +2194,7 @@ export default makePrefab('Crud with dialogs', attrs, beforeCreate, [
                                               value: false,
                                             },
                                           ),
-                                          type: font('Font', {
+                                          type: font('Text style', {
                                             value: ['Title4'],
                                           }),
                                         },
@@ -2664,7 +2664,7 @@ export default makePrefab('Crud with dialogs', attrs, beforeCreate, [
                                                                     },
                                                                   ),
                                                                 type: font(
-                                                                  'Font',
+                                                                  'Text style',
                                                                   {
                                                                     value: [
                                                                       'Title4',
@@ -3013,7 +3013,7 @@ export default makePrefab('Crud with dialogs', attrs, beforeCreate, [
                                                                     },
                                                                   ),
                                                                 type: font(
-                                                                  'Font',
+                                                                  'Text style',
                                                                   {
                                                                     value: [
                                                                       'Title4',
@@ -3447,7 +3447,7 @@ export default makePrefab('Crud with dialogs', attrs, beforeCreate, [
                                                                     },
                                                                   ),
                                                                 type: font(
-                                                                  'Font',
+                                                                  'Text style',
                                                                   {
                                                                     value: [
                                                                       'Title4',
@@ -3881,7 +3881,7 @@ export default makePrefab('Crud with dialogs', attrs, beforeCreate, [
                                                                     },
                                                                   ),
                                                                 type: font(
-                                                                  'Font',
+                                                                  'Text style',
                                                                   {
                                                                     value: [
                                                                       'Title4',
@@ -4333,7 +4333,9 @@ export default makePrefab('Crud with dialogs', attrs, beforeCreate, [
                                         ],
                                       },
                                     }),
-                                    type: font('Font', { value: ['Body1'] }),
+                                    type: font('Text style', {
+                                      value: ['Body1'],
+                                    }),
                                     styles: toggle('Styles', { value: true }),
                                     textColor: color('Text color', {
                                       value: ThemeColor.MEDIUM,