+++ src/prefabs/pageWithCrudSlideOutPanel.tsx
@@ -865,7 +865,7 @@ const drawerContainer = DrawerContainer(
                                                           },
                                                         },
                                                       ),
-                                                      type: font('Font', {
+                                                      type: font('Text style', {
                                                         value: ['Title4'],
                                                       }),
                                                       textColor: color(
@@ -1000,7 +1000,7 @@ const drawerContainer = DrawerContainer(
                                                                     },
                                                                   ),
                                                                 type: font(
-                                                                  'Font',
+                                                                  'Text style',
                                                                   {
                                                                     value: [
                                                                       'Title4',
@@ -1129,7 +1129,7 @@ const drawerContainer = DrawerContainer(
                                                                     },
                                                                   ),
                                                                 type: font(
-                                                                  'Font',
+                                                                  'Text style',
                                                                   {
                                                                     value: [
                                                                       'Body1',
@@ -1462,7 +1462,9 @@ const drawerContainer = DrawerContainer(
                                       ],
                                     },
                                   }),
-                                  type: font('Font', { value: ['Body1'] }),
+                                  type: font('Text style', {
+                                    value: ['Body1'],
+                                  }),
                                   styles: toggle('Styles', { value: true }),
                                   textColor: color('Text color', {
                                     value: ThemeColor.MEDIUM,
@@ -1639,7 +1641,7 @@ const drawerBar = DrawerBar(
                                     value: ['Create'],
                                     configuration: { as: 'MULTILINE' },
                                   }),
-                                  type: font('Font', {
+                                  type: font('Text style', {
                                     value: ['Title5'],
                                   }),
                                   textColor: color('Text color', {
@@ -1947,7 +1949,7 @@ const drawerBar = DrawerBar(
                                     value: ['Details'],
                                     configuration: { as: 'MULTILINE' },
                                   }),
-                                  type: font('Font', {
+                                  type: font('Text style', {
                                     value: ['Title5'],
                                   }),
                                   textColor: color('Text color', {
@@ -2241,7 +2243,7 @@ const drawerBar = DrawerBar(
                                     value: ['Update'],
                                     configuration: { as: 'MULTILINE' },
                                   }),
-                                  type: font('Font', {
+                                  type: font('Text style', {
                                     value: ['Title5'],
                                   }),
                                   textColor: color('Text color', {