+++ src/prefabs/pageWithRegisterForm.tsx
@@ -977,7 +977,7 @@ export default makePrefab('Register form', attrs, beforeCreate, [
                                                   as: 'MULTILINE',
                                                 },
                                               }),
-                                              type: font('Font', {
+                                              type: font('Text style', {
                                                 value: ['Title4'],
                                               }),
                                             },